chore(gmeter): convert to uint64

+++ meter/mem.go
@@ -4,5 +4,5 @@ package meter
 import "C"
 
 func Mem() uint64 {
-	return C.sysconf(C._SC_PHYS_PAGES) * C.sysconf(C._SC_PAGE_SIZE)
+	return uint64(C.sysconf(C._SC_PHYS_PAGES) * C.sysconf(C._SC_PAGE_SIZE))
 }
